Konvertering mellon long og string for organisasjonsnummer foregår i dao

+++ src/main/kotlin/no/nav/helse/modell/arbeidsgiver/ArbeidsgiverLøsning.kt
@@ -2,5 +2,5 @@ package no.nav.helse.modell.arbeidsgiver
 
 internal class ArbeidsgiverLøsning(internal val navn: String) {
     internal fun oppdater(arbeidsgiverDao: ArbeidsgiverDao, orgnummer: String) =
-        arbeidsgiverDao.updateNavn(orgnummer.toLong(), navn)
+        arbeidsgiverDao.updateNavn(orgnummer, navn)
 }
