+++ src/test/kotlin/no/nav/helse/modell/command/nyny/OppdaterArbeidsgiverCommandTest.kt
@@ -30,14 +30,14 @@ internal class OppdaterArbeidsgiverCommandTest {
 
     @Test
     fun `oppdaterer ikke når informasjonen er ny`() {
-        every { dao.findNavnSistOppdatert(ORGNR.toLong()) } returns LocalDate.now()
+        every { dao.findNavnSistOppdatert(ORGNR) } returns LocalDate.now()
         assertTrue(command.execute(context))
         verify(exactly = 0) { dao.updateNavn(any(), any()) }
     }
 
     @Test
     fun `oppdaterer ikke når informasjonen er gammel`() {
-        every { dao.findNavnSistOppdatert(ORGNR.toLong()) } returns LocalDate.now().minusYears(1)
+        every { dao.findNavnSistOppdatert(ORGNR) } returns LocalDate.now().minusYears(1)
         assertTrue(command.execute(context))
         assertFalse(context.harBehov())
         verify(exactly = 0) { dao.updateNavn(any(), any()) }
